冯豆 增加订单验证

+++ mall/src/main/java/com/cn/wanxi/mall/controller/order/OrderController.java
@@ -1,9 +1,14 @@
 package com.cn.wanxi.mall.controller.order;
 
+
 import com.cn.wanxi.entity.order.OrderEntity;
+import com.cn.wanxi.entity.order.PageMap;
 import com.cn.wanxi.utils.utils.Msg;
 import com.cn.wanxi.service.order.IOrderService;
+import io.swagger.models.auth.In;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.util.ObjectUtils;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -30,14 +35,16 @@ public class OrderController {
      */
     @PostMapping(value = "/add", produces = "application/json;charset=UTF-8")
     public Msg add(@RequestBody OrderEntity orderEntity) {
-        Msg m;
-        int result = iOrderService.add(orderEntity);
-        if (!isEmpty(result)) {
-            m = Msg.success().messageData(orderEntity);
+        Msg msg=null;
+        if (null != orderEntity.getUsername() && orderEntity.getUsername().trim() != "") {
+            int result = iOrderService.add(orderEntity);
+            if (0 != result) {
+                msg = Msg.success().messageData(orderEntity);
+            }
         } else {
-            m = Msg.fail();
+            msg = Msg.fail().messageData("名字不能为空");
         }
-        return m;
+        return msg;
     }
 
     @PostMapping(value = "/findPage", produces = "application/json;charset=UTF-8")
@@ -45,42 +52,80 @@ public class OrderController {
         Msg msg = null;
         int page = param.get("page");
         int size = param.get("size");
+        PageMap pageMap = new PageMap();
+        //处理int类型的变量为空时
+        if (page == 0) {
+            page = 1;
+        }
+        if (size == 0) {
+            page = 30;
+        }
 
         Map<String, Object> list = iOrderService.list(page, size);
-        //判断集合是否有数据，如果没有数据返回失败
-        if (list.isEmpty()) {
-            msg = Msg.fail();
+        //把查询出来的对象封装在分页实体类中
+        pageMap.setMap(list);
+        if (null == list && list.isEmpty()) {
+            msg = Msg.fail().messageData("订单信息不存在");
         } else {
-            msg = Msg.success().messageData(list);
+
+            //统计所有数据的总行数
+            int TotalRows = iOrderService.countAll();
+
+            //把页数封装在分页实体类中
+            pageMap.setPage(page);
+            pageMap.setTotal(list.size());
+            //查询出来的总行数封装在分页实体类中
+            pageMap.setTotalRows(TotalRows);
+            msg = getPages(size, pageMap, TotalRows);
         }
         return msg;
     }
 
     @PostMapping(value = "/batchFindAll", produces = "application/json;charset=UTF-8")
-    public Map<String, Object> list2(@RequestBody Map<String, String> param, HttpServletResponse response) {
+    public Msg list2(@RequestBody Map<String, String> param) {
         Msg msg = null;
-        String ids=param.get("ids");
-        Map<String, Object> list = iOrderService.batchlist(ids);
-        //判断集合是否有数据，如果没有数据返回失败
-        if (list.isEmpty()) {
-            msg = Msg.fail();
+        String ids = param.get("ids");
+        if("".equals(ids)){
+            msg = Msg.fail().messageData("格式不对");
         } else {
-            msg = Msg.success().messageData(list);
+            Map<String, Object> list = iOrderService.batchlist(ids);
+
+            //判断集合是否有数据，如果没有数据返回失败
+            if (!ObjectUtils.isEmpty(list)) {
+                msg = Msg.success().messageData(list);
+            } else {
+                msg = Msg.fail().messageData("不存在");
+            }
         }
-        return list;
+
+        return msg;
     }
 
     @PostMapping(value = "/batchSendSubmit", produces = "application/json;charset=UTF-8")
-    public Msg batchSendSubmit(@RequestBody Map<String, Integer> param, HttpServletResponse response) {
-        Msg m;
-        int orderId = param.get("orderId");
-        int result = iOrderService.batchSendSubmit(orderId);
-        if (!isEmpty(result)) {
-            m = Msg.success().messageData(orderId);
+    public Msg batchSendSubmit(@RequestBody Map<String, String> param) {
+        Msg msg = null;
+
+        Object shippingName = param.get("shippingName");
+        Object shippingCode = param.get("shippingCode");
+        Object orderId = param.get("orderId");
+        int id = Integer.parseInt(String.valueOf(orderId));
+        if (id > 0) {
+            //根据id查询数据
+            OrderEntity byId = iOrderService.findById(id);
+            //判断是否查询到该品牌信息
+            if (!ObjectUtils.isEmpty(byId)) {
+                int result = iOrderService.batchSendSubmit(id, orderId, shippingName, shippingCode);
+                if (result > 0) {
+                    msg = Msg.success();
+                }
+            } else {
+                msg = Msg.fail().messageData("该订单不存在");
+            }
         } else {
-            m = Msg.fail();
+            msg = Msg.fail().messageData("输入格式有误");
         }
-        return m;
+        return msg;
+
     }
 //    @PostMapping("/findbyorderid")
 //    public Msg findByOrderId(int orderId) {
@@ -94,16 +139,24 @@ public class OrderController {
 //        return msg;
 //    }
 
-
+    /**
+     * 根据订单主表id查询信息
+     *
+     * @param param
+     * @return
+     */
     @PostMapping(value = "/findbyid", produces = "application/json;charset=UTF-8")
-    public Msg findById(@RequestBody Map<String, Integer> param, HttpServletResponse response) {
+    public Msg findById(@RequestBody Map<String, Integer> param) {
         Msg msg = null;
         int id = param.get("id");
-        OrderEntity byId = iOrderService.findById(id);
-        if (byId != null) {
-            msg = Msg.success().messageData(byId);
-        } else {
-            msg = Msg.fail();
+        if (!StringUtils.isEmpty(id) && id > 0) {
+            OrderEntity byId = iOrderService.findById(id);
+            //判断是否有返回的数据
+            if (!ObjectUtils.isEmpty(byId)) {
+                msg = Msg.success().messageData(byId);
+            } else {
+                msg = Msg.fail().messageData("该订单不存在");
+            }
         }
         return msg;
     }
@@ -111,19 +164,22 @@ public class OrderController {
     /**
      * 【根据id删除】
      *
-     * @param param
-     * @param response
+     * @param orderEntity
      * @return
      */
     @PostMapping(value = "/delete", produces = "application/json;charset=UTF-8")
-    public Msg deleteById(@RequestBody Map<String, Integer> param, HttpServletResponse response) {
+    public Msg deleteById(@RequestBody OrderEntity orderEntity) {
         Msg msg = null;
-        int id = param.get("id");
-        int i = iOrderService.deleteById(id);
-        if (i > 0) {
-            msg = Msg.success();
+        int id = orderEntity.getId();
+        if (id > 0) {
+            int i = iOrderService.deleteById(id);
+            if (i > 0) {
+                msg = Msg.success().messageData("删除成功");
+            } else {
+                msg = Msg.fail().messageData("删除失败,该用户不存在");
+            }
         } else {
-            msg = Msg.fail();
+            msg = Msg.fail().messageData("请输入id");
         }
         return msg;
     }
@@ -146,4 +202,48 @@ public class OrderController {
         }
         return msg;
     }
+
+    /**
+     * 提取公共方法
+     *
+     * @param size
+     * @param pageMap
+     * @param totalRows 总记录数
+     * @return
+     */
+    private Msg getPages(int size, PageMap pageMap, int totalRows) {
+        Msg msg;
+        int pages = 0;
+        if (totalRows % size == 0) {
+            pages = totalRows / size;
+        } else {
+            pages = totalRows / size + 1;
+        }
+        System.out.println("目前分页的总页数是" + pages);
+        //总页数
+        pageMap.setPages(pages);
+        msg = Msg.success().messageData(pageMap);
+        return msg;
+    }
+
+//    @PostMapping(value = "/test", produces = "application/json;charset=UTF-8")
+//    public Map<String, Object> test(@RequestBody Map<String, String> param) {
+//        Msg msg = null;
+//        String ids = param.get("ids");
+//        if ("".equals(ids)) {
+//            msg = Msg.fail().messageData("aaa");
+//        } else {
+//            ids = "aa,sss,,ddds,io";
+//            String[] idList = ids.split(",");
+//            Integer.parseInt(idList[0]);
+//            Map<String, Object> list = iOrderService.batchlist(ids);
+//            //判断集合是否有数据，如果没有数据返回失败
+//            if (list.isEmpty()) {
+//                msg = Msg.fail();
+//            } else {
+//                msg = Msg.success().messageData(list);
+//            }
+//        }
+//        return null;
+//    }
 }
